requirements.txt
customtkinter==5.2.2
tkinterweb==4.3.1
markdown==3.4.1
openai==1.75.0
python-dotenv==1.1.0
PyPDF2==3.0.1
requests==2.32.3
playwright==1.50.0
psutil==5.9.8
pywin32==307
reportlab==4.3.1
pandas==2.2.3
